requirements.txt
beautifulsoup4==4.12.2
Jinja2==3.1.2
selenium==3.141.0
selenium_wire==1.0.8
lxml==4.9.2
requests==2.28.2
webdriver-manager==4.0.0
minify_html==0.15.0
rfeed==1.1.1
Unidecode==1.3.6
Pillow==9.0.1
